chore(lts): move group and stream into service dir
